# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SocialPredict API
  version: 0.1.0
  description: >
    HTTP contract for the SocialPredict backend service. This spec is generated
    from the Go handlers and DTOs under backend/server and backend/handlers.
    Routes wrapped by the shared security middleware can return a shared
    transport-level `429` JSON failure envelope response before reaching handler-specific
    JSON envelope logic when the rate limit is exceeded. This middleware
    exception applies across protected routes even when a path omits a repeated
    path-level `429` entry.
servers:
  - url: http://localhost:8080
    description: Local dev server
...
paths:
...
  /v0/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: Authenticate user
      description: >
        Validates username and password and returns a JWT bearer token for subsequent
        authenticated requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid JSON payload or validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '401':
          description: Invalid username or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '500':
          description: Database access or token creation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          description: Login rate limit exceeded by middleware. Other protected routes can also return middleware-level 429 responses as described in the API description.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
...
  /v0/setup:
    get:
      tags: [Config]
      operationId: getEconomicsConfig
      summary: Get economics configuration
      description: Returns the current economic configuration used by the backend.
      responses:
        '200':
          description: Economics configuration returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Economics'
        '500':
          description: Failed to load or encode configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          description: Rate limit exceeded by shared security middleware.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
  /v0/setup/frontend:
    get:
      tags: [Config]
      operationId: getFrontendConfig
      summary: Get frontend configuration
      description: Returns minimal configuration values needed by the frontend, such as chart precision.
      responses:
        '200':
          description: Frontend configuration returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrontendConfig'
        '500':
          description: Failed to load or encode configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          description: Rate limit exceeded by shared security middleware.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
...
  /v0/markets:
    get:
      tags: [Markets]
      operationId: listMarkets
      summary: List markets
      description: >
        Returns markets with optional filtering by status and creator. When status is
        provided, results are backed by a status-specific query.
      parameters:
        - in: query
          name: status
          required: false
          description: >
            Filter by market status. Accepted values are active, closed, resolved, and all
            (or omitted for all markets).
          schema:
            type: string
            enum: [active, closed, resolved, all]
        - in: query
          name: created_by
          required: false
          description: Only include markets created by the given username.
          schema:
            type: string
        - in: query
          name: limit
          required: false
          description: Maximum number of markets to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          required: false
          description: Number of markets to skip before collecting results.
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '400':
          description: Invalid status filter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '500':
          description: Unexpected server error while listing markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          description: Rate limit exceeded by shared security middleware.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
    post:
      tags: [Markets]
      operationId: createMarket
      summary: Create a market
      description: Creates a new prediction market for the authenticated user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMarketRequest'
      responses:
        '201':
          description: Market created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketResponse'
        '400':
          description: Invalid request body or market validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '401':
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '403':
          description: Password change required before creating markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '404':
          description: Authenticated user could not be loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '422':
          description: Creator balance does not allow market creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '500':
          description: Unexpected server error during market creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
  /v0/markets/search:
    get:
      tags: [Markets]
      operationId: searchMarkets
      summary: Search markets
      description: >
        Searches markets by question title with optional status filtering. Supports both
        "query" and legacy "q" parameters.
      parameters:
        - in: query
          name: query
          required: false
          description: Search query (preferred parameter name).
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - in: query
          name: q
          required: false
          description: Legacy search query parameter (used if `query` is not provided).
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - in: query
          name: status
          required: false
          description: Filter by market status; same accepted values as /v0/markets.
          schema:
            type: string
            enum: [active, closed, resolved, all]
        - in: query
          name: limit
          required: false
          description: Maximum number of primary results to return.
          schema:
            type: integer
            minimum: 1
            maximum: 50
        - in: query
          name: offset
          required: false
          description: Number of primary results to skip before collecting.
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Search completed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Missing query parameter or invalid status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '500':
          description: Unexpected server error during search.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          description: Rate limit exceeded by shared security middleware.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
  /v0/markets/status:
    get:
      tags: [Markets]
      operationId: listAllMarketStatuses
      summary: List markets across all statuses
      description: Convenience endpoint equivalent to `/v0/markets?status=all`.
      parameters:
        - in: query
          name: limit
          required: false
          description: Maximum number of markets to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          required: false
          description: Number of markets to skip before collecting results.
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '500':
          description: Unexpected server error while listing markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          description: Rate limit exceeded by shared security middleware.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
  /v0/markets/status/{status}:
    get:
      tags: [Markets]
      operationId: listMarketsByStatus
      summary: List markets by status
      description: Returns markets filtered by a specific lifecycle status.
      parameters:
        - in: path
          name: status
          required: true
          description: >
            Market status to filter by. Accepted values are active, closed, resolved, and all.
          schema:
            type: string
            enum: [active, closed, resolved, all]
        - in: query
          name: limit
          required: false
          description: Maximum number of markets to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          required: false
          description: Number of markets to skip before collecting results.
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '400':
          description: Invalid or unsupported status value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '500':
          description: Unexpected server error while listing markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          description: Rate limit exceeded by shared security middleware.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
  /v0/markets/active:
    get:
      tags: [Markets]
      operationId: listActiveMarkets
      summary: List active markets
      description: Legacy convenience endpoint equivalent to `/v0/markets?status=active`.
      parameters:
        - in: query
          name: limit
          required: false
          description: Maximum number of markets to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          required: false
          description: Number of markets to skip before collecting results.
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '500':
          description: Unexpected server error while listing markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          description: Rate limit exceeded by shared security middleware.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
  /v0/markets/closed:
    get:
      tags: [Markets]
      operationId: listClosedMarkets
      summary: List closed markets
      description: Legacy convenience endpoint equivalent to `/v0/markets?status=closed`.
      parameters:
        - in: query
          name: limit
          required: false
          description: Maximum number of markets to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          required: false
          description: Number of markets to skip before collecting results.
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '500':
          description: Unexpected server error while listing markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          description: Rate limit exceeded by shared security middleware.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
  /v0/markets/resolved:
    get:
      tags: [Markets]
      operationId: listResolvedMarkets
      summary: List resolved markets
      description: Legacy convenience endpoint equivalent to `/v0/markets?status=resolved`.
      parameters:
        - in: query
          name: limit
          required: false
          description: Maximum number of markets to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          required: false
          description: Number of markets to skip before collecting results.
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '500':
          description: Unexpected server error while listing markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          description: Rate limit exceeded by shared security middleware.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
    LoginResponse:
      type: object
      required: [ok, result]
      properties:
        ok:
          type: boolean
          enum: [true]
        result:
          $ref: '#/components/schemas/LoginResult'
...
    ListMarketsResponse:
      type: object
      properties:
        markets:
          type: array
          items:
            $ref: '#/components/schemas/MarketOverviewResponse'
        total:
          type: integer
...
    Economics:
      type: object
      required: [marketcreation, marketincentives, user, betting]
      properties:
        marketcreation:
          $ref: '#/components/schemas/MarketCreation'
        marketincentives:
          $ref: '#/components/schemas/MarketIncentives'
        user:
          $ref: '#/components/schemas/UserEconomics'
        betting:
          $ref: '#/components/schemas/Betting'
...
    FrontendConfig:
      type: object
      required: [charts]
      properties:
        charts:
          $ref: '#/components/schemas/FrontendCharts'
...